backend/requirements.txt
Django
djangorestframework
django-cors-headers
python-decouple
django-mongodb-backend
pymongo
